requirements.txt
astroid==2.4.2
boto3==1.12.41
botocore==1.15.49
certifi==2020.11.8
chardet==3.0.4
Django==2.1
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyfcm==1.5.1
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
sendgrid==6.4.7
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
wrapt==1.12.1
stripe==2.56.0
twilio==6.58.0
django-sqlprint-middleware==0.1.3
pygments-pprint-sql==0.1.0
Pygments==2.9.0